+++ cassandra/cluster.py
@@ -1629,6 +1629,7 @@ class ControlConnection(object):
             if partitioner and tokens:
                 token_map[host] = tokens
 
+        should_rebuild_token_map = False
         found_hosts = set()
         for row in peers_result:
             addr = row.get("rpc_address")
@@ -1645,8 +1646,9 @@ class ControlConnection(object):
             if host is None:
                 log.debug("[control connection] Found new host to connect to: %s", addr)
                 host = self._cluster.add_host(addr, datacenter, rack, signal=True)
+                should_rebuild_token_map = True
             else:
-                self._update_location_info(host, datacenter, rack)
+                should_rebuild_token_map |= self._update_location_info(host, datacenter, rack)
 
             tokens = row.get("tokens")
             if partitioner and tokens:
@@ -1657,15 +1659,17 @@ class ControlConnection(object):
                     old_host.address not in found_hosts and \
                     old_host.address not in self._cluster.contact_points:
                 log.debug("[control connection] Found host that has been removed: %r", old_host)
+                should_rebuild_token_map = True
                 self._cluster.remove_host(old_host)
 
-        if partitioner:
-            log.debug("[control connection] Fetched ring info, rebuilding metadata")
+        log.debug("[control connection] Finished fetching ring info")
+        if partitioner and should_rebuild_token_map:
+            log.debug("[control connection] Rebuilding token map due to topology changes")
             self._cluster.metadata.rebuild_token_map(partitioner, token_map)
 
     def _update_location_info(self, host, datacenter, rack):
         if host.datacenter == datacenter and host.rack == rack:
-            return
+            return False
 
         # If the dc/rack information changes, we need to update the load balancing policy.
         # For that, we remove and re-add the node against the policy. Not the most elegant, and assumes
@@ -1673,6 +1677,7 @@ class ControlConnection(object):
         self._cluster.load_balancing_policy.on_down(host)
         host.set_location_info(datacenter, rack)
         self._cluster.load_balancing_policy.on_up(host)
+        return True
 
     def _handle_topology_change(self, event):
         change_type = event["change_type"]