+++ CHANGELOG.rst
@@ -19,6 +19,8 @@ Bug Fixes
 * Fix unintended rebuild of token replica map when keyspaces are
   discovered (on startup), added, or updated and TokenAwarePolicy is not
   in use.
+* Avoid rebuilding token metadata when cluster topology has not
+  actually changed
 
 1.1.2
 =====